requirements.txt
tornado==5.1.1
Pillow==5.3.0
pytesseract
urllib
